+++ backend/src/main/kotlin/com/otoki/powersales/claim/dto/response/ClaimDetailResponse.kt
@@ -4,6 +4,7 @@ import com.otoki.powersales.claim.entity.Claim
 import com.otoki.powersales.claim.entity.ClaimPhoto
 import java.time.LocalDate
 import java.time.LocalDateTime
+import java.math.BigDecimal
 
 data class ClaimDetailResponse(
     val claimId: Long,
@@ -18,8 +19,8 @@ data class ClaimDetailResponse(
     val subcategoryValue: String?,
     val subcategoryLabel: String?,
     val defectDescription: String?,
-    val defectQuantity: Long?,
-    val purchaseAmount: Long?,
+    val defectQuantity: BigDecimal?,
+    val purchaseAmount: BigDecimal?,
     val purchaseMethodName: String?,
     val requestTypeName: String?,
     val status: String,
