+++ backend/src/main/kotlin/com/otoki/powersales/promotion/dto/request/BatchUpdatePromotionEmployeeRequest.kt
@@ -6,6 +6,7 @@ import jakarta.validation.constraints.NotNull
 import jakarta.validation.constraints.Min
 import jakarta.validation.constraints.Size
 import java.time.LocalDate
+import java.math.BigDecimal
 
 data class BatchUpdatePromotionEmployeeRequest(
     @field:NotEmpty(message = "items는 1건 이상이어야 합니다")
@@ -31,10 +32,10 @@ data class BatchUpdatePromotionEmployeeItem(
     val workType3: String? = null,
 
     @field:Min(value = 0, message = "판매단가는 0 이상이어야 합니다")
-    val basePrice: Long? = null,
+    val basePrice: BigDecimal? = null,
 
     @field:Min(value = 0, message = "일일 목표수량은 0 이상이어야 합니다")
-    val dailyTargetCount: Long? = null,
+    val dailyTargetCount: BigDecimal? = null,
 
     @field:Min(value = 0, message = "목표금액은 0 이상이어야 합니다")
     val targetAmount: Long? = 0,
@@ -43,19 +44,19 @@ data class BatchUpdatePromotionEmployeeItem(
     val actualAmount: Long? = 0,
 
     @field:Min(value = 0, message = "대표품목 매출금액은 0 이상이어야 합니다")
-    val primaryProductAmount: Long? = null,
+    val primaryProductAmount: BigDecimal? = null,
 
     @field:Min(value = 0, message = "대표품목 판매수량은 0 이상이어야 합니다")
-    val primarySalesQuantity: Long? = null,
+    val primarySalesQuantity: BigDecimal? = null,
 
     @field:Min(value = 0, message = "대표품목 판매단가는 0 이상이어야 합니다")
-    val primarySalesPrice: Long? = null,
+    val primarySalesPrice: BigDecimal? = null,
 
     @field:Min(value = 0, message = "기타 판매금액은 0 이상이어야 합니다")
-    val otherSalesAmount: Long? = null,
+    val otherSalesAmount: BigDecimal? = null,
 
     @field:Min(value = 0, message = "기타 판매수량은 0 이상이어야 합니다")
-    val otherSalesQuantity: Long? = null,
+    val otherSalesQuantity: BigDecimal? = null,
 
     @field:Size(max = 255, message = "현장사진 S3 키는 최대 255자입니다")
     val s3ImageUniqueKey: String? = null
