+++ backend/src/test/kotlin/com/otoki/powersales/claim/service/ClaimQueryServiceTest.kt
@@ -28,6 +28,7 @@ import org.mockito.kotlin.whenever
 import java.time.LocalDate
 import java.time.LocalDateTime
 import java.util.*
+import java.math.BigDecimal
 
 @ExtendWith(MockitoExtension::class)
 @DisplayName("ClaimQueryService 테스트")
@@ -189,7 +190,7 @@ class ClaimQueryServiceTest {
     private fun createClaim(
         id: Long = 1L,
         employeeId: Long = 1L,
-        purchaseAmount: Long? = 3500L,
+        purchaseAmount: BigDecimal? = BigDecimal.valueOf(3500L),
         purchaseMethodName: String? = "개인카드",
         requestTypeName: String? = "교환"
     ): Claim {
@@ -205,7 +206,7 @@ class ClaimQueryServiceTest {
             claimType1 = ClaimType1.A,
             claimType2 = ClaimType2.AB,
             defectDescription = "포장 파손으로 내용물 누유 확인",
-            defectQuantity = 3L,
+            defectQuantity = BigDecimal.valueOf(3L),
             purchaseAmount = purchaseAmount,
             purchaseMethodName = purchaseMethodName,
             requestTypeName = requestTypeName,
